The process of ordering coffee is adjusted.

+++ test/src/main/java/com/scand/test/repositories/CoffeeOrderRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CoffeeOrderRepository extends JpaRepository<CoffeeOrder, Long>
+public interface CoffeeOrderRepository extends JpaRepository<CoffeeOrder, Integer>
 {
 }
